+++ src/mpi/apps/tbsla_personalized_page_rank_mpi.cpp
@@ -113,9 +113,9 @@ int main(int argc, char **argv){
   b = m->personalized_page_rank(MPI_COMM_WORLD, beta, epsilon, max_iterations, personalized_nodes, nb_iterations_done);
   auto t_personalized_page_rank_end = now();
 
-  int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
-  int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
-  int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
+  long int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
+  long int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
+  long int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
   
   if(rank == 0){
     auto t_app_end = now();
